break apart BuiltInFunctions mega class
